Housekeeping tasks to get gem release ready
We now have this awesome gem! Our normal pattern is to actually release it as a gem to [rubygems](https://rubygems.org/) and then reference it from there in the projects that use it (at the moment we are referencing the GitHub repos directly).

Before we do that though, there are a few housekeeping tasks we need to perform to the project. Some quick examples

- test coverage is not actually setup
- license is currently wrong
- we've committed test db's to the project which get changed on each test run

So this change is a general smartening up of the repo before we cut a 0.1.0 release.