Move dependencies from Gemfile to .gemspec
As the instructions in the Gemfile say, and as per Yehuda Katz's post [Clarifying the Roles of the .gemspec and Gemfile](https://yehudakatz.com/2010/12/16/clarifying-the-roles-of-the-gemspec-and-gemfile/) our gem's dependencies should be declared in the .gemspec. The Gemfile should be reserved for experimental or temporary dependencies.

This is the model we have followed in our other gems so updating this one to match. It also highlighted we were getting Devise and CanCanCan autoloaded for free because of their inclusion on the Gemfile. Hence the need to require them specifically in the dummy's `config/application.rb`.