Remove unused spec
This appears to be something generated that we are not currently using. So to cimplify things this change removes it (not sure what its doing in the dummy folder either).

Also had to add the `require "rails_helper"` to the config spec. It looks like we were reliant on it being called in the unused spec we're deleting in order for the tests to run.