Move pry in rspec to support file
Our pattern is all additional tools or dependencies needed for testing we handle in separate support files. That way we find it easier to ensure our spec and rails helper files stay consistent across projects. It also makes it clearer and easier to see just what additional tools a test suite is using.