Housekeeping tasks to get gem release ready (#2)
We now have this awesome gem! Our normal pattern is to actually release it as a gem to [rubygems](https://rubygems.org/) and then reference it from there in the projects that use it (at the moment we are referencing the GitHub repos directly).

Before we do that though, there are a few housekeeping tasks we need to perform to the project. Some quick examples

- test coverage is not actually setup
- license is currently wrong
- we've committed test db's to the project which get changed on each test run

So this change is a general smartening up of the repo before we cut a 0.1.0 release.

- Sort out LICENSE
- Update .gitignore
- Move dependencies from Gemfile to .gemspec
- Remove unused spec
- Move requires to engine
- Fix broken test coverage reporting
- Move pry in rspec to support file
- Remove factory-bot comments
- Update comments to match other rails_helper files
- Move migrations checker to support file
- Remove unused fixture config
- Update Rspec config to match other projects
- Remove RSpec warnings config
- Remove tasks folder
- Tidy up requires in bin and dummy app
- Remove some unused files
- Remove spec_helper comment
- Add something to config test
- Add support for rspec test generators