Update reference to Cancancan in gemspec
https://eaflood.atlassian.net/browse/RUBY-1139

We'd like the ability to switch off features in our services from the UI rather than having to go through an RfC process. Because we want to do this in multiple services we built this engine.

This change is dealing with an issue that has arisen between WCR and WEX. WCR uses MongoDb and hence the [Mongoid framework](https://github.com/mongodb/mongoid). WEX uses PostgreSql. Both use [cancancan](https://github.com/CanCanCommunity/cancancan) for authorzation.

However, support for Mongoid in CanCanCan was dropped from v2 onwards so WCR has CanCanCan locked to v1.10. WEX though expects v2.3 or greater. Because of this we are changing how CanCanCan is referenced in the spec. Rather than saying v1.10.x is expected we now just say v1.10 is the minimum.